zanata_users: Consider coordinators as translators/reviewers
In Zanata 3.9.6 language team panel, language coordinators
are just shown as 'Coordinator'.
Previously 'Translator' and 'Reviewer' are shown as their role.
zanata_users.py script depends on what is displayed.

'translators' field in translation_team.yaml is used
in Stackalytics and zanata_stats.py. In the current situation,
activities in language coordinators are not retrieved.

Language coordinators can also be a translator and a reviewer
of corresponding language team, so this commit changes
zanata_users.py to add coordinators to translators and reviewers
of corresponding language.

Change-Id: Ia137998edb97dc45c576505a11e84d8e62000e81