Use the new html_context of openstackdocstheme
The new way of configuring html_context was introduced
in openstackdocstheme 1.11.0.
It is much simpler and most projects adopt it now.

Note that the current openstackdocstheme does not declare if it is
safe for parallel reading, so -j 2 or larget int does not work.
Let's drop -j option as the doc build is fast enough.

Change-Id: If1918b0101abf93aa55d4abe3c91a13bb0fe83d1