Rename glossary files to more meaningful names
The directory and filename were decided based on how the infra
translation script handles it, but the glossary files are not
synced with Zanata automatically and manually maintained.

It looks better to use more meaningful file names.

Even if the glossary files are synced in future, the syncing process
would be different from usual trnaslation sync, so I believe
changing the file path would be no problem.

Change-Id: Ib00e0cab46e2ba624eaf30d2fe3127979090c444