Add a check if a proposed team member list is current
It is a boring job to review a patch to update the team member list.
To avoid this, this commit proposes a bit improved approach.

- When you want to propose an update of the team member list,
  you need to sync all members with Zanata.
  If you want to add you, you need to propose a latest member list.
  The new list may contains someone other than you.
- pep8 job checks a proposed member list is up-to-date.
  If not, the job fails.

To make it easy to sync all members with Zanata, a tox target
'zanata-users-sync' is added. To download the latest member list
of all language teams, just run:

  tox -e zanata-users-sync

Change-Id: I393a6b77261f6a3d7788a4c4f57bd6fdc95b5146