Sort entries in translation_team.yaml alphabetically
Previously we kept the order of members of individual language teams
in the same order as that of Zanata language team page.
It was mainly to make it the maintenance of the YAML file
as the file is manually maintained.

We use zanata_users.py script to sync translation_team.yaml now,
so the need to keep the order has decreased much.
In addition, it seems Zanata changes the order of members
even when members of a specific language team is not changed.
This happens during the review https://review.openstack.org/#/c/485966/.

This commit change the script to sort members in the alphabetical order.

Change-Id: I09ddc202e0a3021bbc5433e37a5d815c760697ff