Add special docs build job
The new PTI doesn't work for the way the i18n docs are built. Add a new
doc build job that runs tox -edocs for until we figure out a better path
to success.

Remove the add-marker code since that's done in the doc publication
jobs. Also stop rsyncing the code to a different directory, since it'll
make the normal build case wrong and is otherwise handled by the
publication jobs.

Change-Id: I5237e750b4bc4899d71952a14c8d2ca92d9985a0