doc: avoid RST files in subdirectory
We are discussing an approach on translation support in
project documents in https://review.openstack.org/#/c/543220/.
In this approach POT files for RST files in subdirectory are
treated differently. In the i18n doc, atc-stats directory is
the only exception and the thing would be much simpler if it
is moved to a top directory.

tools/generatepot.sh is updated accordingly.

Also generatepot.sh is added to pep8 tox env to confirm
generatepot.sh works successfully.

Change-Id: Ia227358cf9681e8d1745f4c260fe9c97ab18a3e7