Fix doc-pot-filter to exclude ATC stats
The path of ATC stats CSV files in the generated POT seems to have changed.
As a result, strings from ATC stats are included to the generated POT file.
There is no need to translate them and the translation percentage will go
down. This commit adjust the filepath in doc-pot-filter.sh to make the POT
filter work as expected again.

Change-Id: Ieeee86619a561fcca216b568797bd6c3c1de4d62