generatepot.sh: Drop UUID filtering
(g)awk in Ubuntu 20.04 or later (perhaps gawk >=5.0) has more strict
regexp check and the regexp passed to awk in generatepot.sh cannot
pass the regexp check. This means our script does not work on
Ubuntu 20.04 or later.

This regexp was introduced to strip lines with UUID in a POT file
(e.g., "# 55a7796f27cb4468820fa48b90515add"), but the recent versions
of sphinx does not output such information at least with the sphinx
configuration used in OpenStack CI.
As a result this awk expression in generatepot.sh does nothing,
and I think we can drop this awk expression.

Change-Id: Ia61f74d0a73cc012afde6a4e16ac4431687a65a1